Added empty Growth tab
